requirements.txt
# Core Django
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
python-decouple>=3.8

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
cryptography>=41.0.0
bcrypt>=4.0.0

# API Documentation
drf-spectacular>=0.26.0

# Background Tasks
celery>=5.3.0
redis>=4.5.0
django-redis>=5.3.0

# WebSockets
channels>=4.0.0
channels-redis>=4.1.0

# File Handling
Pillow>=10.0.0
python-magic>=0.4.27

# Development Tools
django-extensions>=3.2.0
ipython>=8.0.0

# Monitoring & Logging
sentry-sdk>=1.28.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Linting & Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0

# Health Checks
django-health-check>=3.17.0

# Two-Factor Authentication
django-otp>=1.1.0
qrcode>=7.4.0

# Email
django-ses>=3.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0